Move cache tests into default_cache
